Calculate elapsed time as double for accuracy.

+++ source/timers.hpp
@@ -19,7 +19,7 @@ inline int64_t time_diff_ms(uint64_t f, uint64_t s) {
 }
 
 inline float time_diff_s(uint64_t f, uint64_t s) {
-	return (float)(s - f) / (float)(1000UL * TB_TIMER_CLOCK);
+	return float(double(s - f) / double(1000UL * TB_TIMER_CLOCK));
 }
 
 #endif
